Terraria 1.3.5.2

+++ Terraria/GameContent/Generation/ShapeFloodFill.cs
@@ -1,7 +1,7 @@
 ﻿// Decompiled with JetBrains decompiler
 // Type: Terraria.GameContent.Generation.ShapeFloodFill
 // Assembly: Terraria, Version=1.3.5.1, Culture=neutral, PublicKeyToken=null
-// MVID: DF0400F4-EE47-4864-BE80-932EDB02D8A6
+// MVID: E90A5A2F-CD10-4A2C-9D2A-6B036D4E8877
 // Assembly location: F:\Steam\steamapps\common\Terraria\Terraria.exe
 
 using Microsoft.Xna.Framework;
@@ -29,24 +29,24 @@ namespace Terraria.GameContent.Generation
       while (pointQueue.Count > 0 && maximumActions > 0)
       {
         Point point = pointQueue.Dequeue();
-        if (!point16Set.Contains(new Point16((int) point.X, (int) point.Y)) && this.UnitApply(action, origin, (int) point.X, (int) point.Y))
+        if (!point16Set.Contains(new Point16(point.X, point.Y)) && this.UnitApply(action, origin, point.X, point.Y))
         {
           point16Set.Add(new Point16(point));
           --maximumActions;
           if (point.X + 1 < Main.maxTilesX - 1)
-            pointQueue.Enqueue(new Point(point.X + 1, (int) point.Y));
+            pointQueue.Enqueue(new Point(point.X + 1, point.Y));
           if (point.X - 1 >= 1)
-            pointQueue.Enqueue(new Point(point.X - 1, (int) point.Y));
+            pointQueue.Enqueue(new Point(point.X - 1, point.Y));
           if (point.Y + 1 < Main.maxTilesY - 1)
-            pointQueue.Enqueue(new Point((int) point.X, point.Y + 1));
+            pointQueue.Enqueue(new Point(point.X, point.Y + 1));
           if (point.Y - 1 >= 1)
-            pointQueue.Enqueue(new Point((int) point.X, point.Y - 1));
+            pointQueue.Enqueue(new Point(point.X, point.Y - 1));
         }
       }
       while (pointQueue.Count > 0)
       {
         Point point = pointQueue.Dequeue();
-        if (!point16Set.Contains(new Point16((int) point.X, (int) point.Y)))
+        if (!point16Set.Contains(new Point16(point.X, point.Y)))
         {
           pointQueue.Enqueue(point);
           break;
