+++ Terraria/RemoteClient.cs
@@ -1,7 +1,7 @@
 ﻿// Decompiled with JetBrains decompiler
 // Type: Terraria.RemoteClient
 // Assembly: Terraria, Version=1.3.5.1, Culture=neutral, PublicKeyToken=null
-// MVID: DF0400F4-EE47-4864-BE80-932EDB02D8A6
+// MVID: E90A5A2F-CD10-4A2C-9D2A-6B036D4E8877
 // Assembly location: F:\Steam\steamapps\common\Terraria\Terraria.exe
 
 using Microsoft.Xna.Framework;
@@ -89,8 +89,8 @@ namespace Terraria
     public static void CheckSection(int playerIndex, Vector2 position, int fluff = 1)
     {
       int index1 = playerIndex;
-      int sectionX = Netplay.GetSectionX((int) (position.X / 16.0));
-      int sectionY1 = Netplay.GetSectionY((int) (position.Y / 16.0));
+      int sectionX = Netplay.GetSectionX((int) ((double) position.X / 16.0));
+      int sectionY1 = Netplay.GetSectionY((int) ((double) position.Y / 16.0));
       int num = 0;
       for (int index2 = sectionX - fluff; index2 < sectionX + fluff + 1; ++index2)
       {
