+++ TerrariaServer/Terraria/GameContent/Generation/TrackGenerator.cs
@@ -1,7 +1,7 @@
 ﻿// Decompiled with JetBrains decompiler
 // Type: Terraria.GameContent.Generation.TrackGenerator
 // Assembly: TerrariaServer, Version=1.3.5.1, Culture=neutral, PublicKeyToken=null
-// MVID: 880A80AC-FC6C-4F43-ABDD-E2472DA66CB5
+// MVID: C2103E81-0935-4BEA-9E98-4159FC80C2BB
 // Assembly location: F:\Steam\steamapps\common\Terraria\TerrariaServer.exe
 
 using Microsoft.Xna.Framework;
@@ -258,7 +258,7 @@ namespace Terraria.GameContent.Generation
 
     public static void Run(Point start)
     {
-      new TrackGenerator().FindPath((int) start.X, (int) start.Y, 250, true);
+      new TrackGenerator().FindPath(start.X, start.Y, 250, true);
     }
 
     private struct TrackHistory
